+++ XUI/XUI/CXImage/CxImage/ximawmf.h
@@ -76,21 +76,21 @@ class CxImageWMF: public CxImage
 
 typedef struct tagRECT16
 {
-	int16_t	left;
-	int16_t	top;
-	int16_t	right;
-	int16_t	bottom;
+	short int	left;
+	short int	top;
+	short int	right;
+	short int	bottom;
 } RECT16;
 
 // taken from Windos 3.11 SDK Documentation (Programmer's Reference Volume 4: Resources)
 typedef struct tagMETAFILEHEADER
 {
-	uint32_t	key;		// always 0x9ac6cdd7
-	uint16_t	reserved1;	// reserved = 0
+	DWORD	key;		// always 0x9ac6cdd7
+	WORD	reserved1;	// reserved = 0
 	RECT16	bbox;		// bounding rectangle in metafile units as defined in "inch"
-	uint16_t	inch;		// number of metafile units per inch (should be < 1440)
-	uint32_t	reserved2;	// reserved = 0
-	uint16_t	checksum;	// sum of the first 10 WORDS (using XOR operator)
+	WORD	inch;		// number of metafile units per inch (should be < 1440)
+	DWORD	reserved2;	// reserved = 0
+	WORD	checksum;	// sum of the first 10 WORDS (using XOR operator)
 } METAFILEHEADER;
 
 #pragma pack()
@@ -98,8 +98,8 @@ typedef struct tagMETAFILEHEADER
 public:
 	CxImageWMF(): CxImage(CXIMAGE_FORMAT_WMF) { }
 
-	bool Decode(CxFile * hFile, int32_t nForceWidth=0, int32_t nForceHeight=0);
-	bool Decode(FILE *hFile, int32_t nForceWidth=0, int32_t nForceHeight=0)
+	bool Decode(CxFile * hFile, long nForceWidth=0, long nForceHeight=0);
+	bool Decode(FILE *hFile, long nForceWidth=0, long nForceHeight=0)
 			{ CxIOFile file(hFile); return Decode(&file,nForceWidth,nForceHeight); }
 
 #if CXIMAGE_SUPPORT_ENCODE
@@ -108,7 +108,7 @@ public:
 #endif // CXIMAGE_SUPPORT_ENCODE
 
 protected:
-	void ShrinkMetafile(int32_t &cx, int32_t &cy);
+	void ShrinkMetafile(int &cx, int &cy);
 	BOOL CheckMetafileHeader(METAFILEHEADER *pmetafileheader);
 	HENHMETAFILE ConvertWmfFiletoEmf(CxFile *pFile, METAFILEHEADER *pmetafileheader);
 	HENHMETAFILE ConvertEmfFiletoEmf(CxFile *pFile, ENHMETAHEADER *pemfh);
