+++ XUI/XUI/CXImage/jasper/jpc/jpc_t1cod.c
@@ -265,17 +265,17 @@ jpc_initmqctxs();
 	}
 
 	for (i = 0; i < (1 << JPC_NMSEDEC_BITS); ++i) {
-		t = i * (float)jpc_pow2i(-JPC_NMSEDEC_FRACBITS);
+		t = i * jpc_pow2i(-JPC_NMSEDEC_FRACBITS);
 		u = t;
-		v = t - 1.5f;
+		v = t - 1.5;
 		jpc_signmsedec[i] = jpc_dbltofix(floor((u * u - v * v) * jpc_pow2i(JPC_NMSEDEC_FRACBITS) + 0.5) / jpc_pow2i(JPC_NMSEDEC_FRACBITS));
 /* XXX - this calc is not correct */
 		jpc_signmsedec0[i] = jpc_dbltofix(floor((u * u) * jpc_pow2i(JPC_NMSEDEC_FRACBITS) + 0.5) / jpc_pow2i(JPC_NMSEDEC_FRACBITS));
-		u = t - 1.0f;
+		u = t - 1.0;
 		if (i & (1 << (JPC_NMSEDEC_BITS - 1))) {
-			v = t - 1.5f;
+			v = t - 1.5;
 		} else {
-			v = t - 0.5f;
+			v = t - 0.5;
 		}
 		jpc_refnmsedec[i] = jpc_dbltofix(floor((u * u - v * v) * jpc_pow2i(JPC_NMSEDEC_FRACBITS) + 0.5) / jpc_pow2i(JPC_NMSEDEC_FRACBITS));
 /* XXX - this calc is not correct */
